% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleuze.R
\name{clr}
\alias{clr}
\title{compute CLR using Aitchison's method}
\usage{
clr(x, cols_as_features = F)
}
\arguments{
\item{x}{A table of compositional data without zeroes.}

\item{cols_as_features}{A boolean. Toggles whether rows or columns are samples.}
}
\value{
A table of CLR-transformed data
}
\description{
CLR transform a table.
}
