% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleuze-package.R
\docType{package}
\name{deleuze-package}
\alias{deleuze}
\alias{deleuze-package}
\title{deleuze: Count data resampling framework}
\description{
Microbiome Oriented count data resampling framework made at the APC microbiome in Ireland.
}
\author{
\strong{Maintainer}: Thomaz Bastiaanssen \email{thomaz.bastiaanssen@ucc.ie}

}
\keyword{internal}
