% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleuze.R
\name{estVolatility}
\alias{estVolatility}
\title{Estimate volatility from count data in Aitchision distance.}
\usage{
estVolatility(count_table, cols_as_features = F, adjust = T)
}
\arguments{
\item{count_table}{A table of count data with rows as features and columns as samples.}

\item{cols_as_features}{A boolean. Toggles whether rows or columns are samples.}
}
\description{
Compute feature-wise variance over time for a group of measurements.
}
