% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{getBetaVars}
\alias{getBetaVars}
\title{Calculate by feature variance from a beta distributions}
\usage{
getBetaVars(count_sample, log_transformed = F)
}
\arguments{
\item{count_sample}{A vector of count data.}

\item{log_transformed}{A boolean, whether to return log-transformed values.}
}
\description{
See rbeta.
}
