% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{getCLRVars}
\alias{getCLRVars}
\title{Calculate by feature variance from a CLR transformed approximation}
\usage{
getCLRVars(count_sample)
}
\arguments{
\item{count_sample}{A vector of count data.}
}
\description{
See rbeta.
}
