% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_params.R
\name{getDeltaCLRMeans}
\alias{getDeltaCLRMeans}
\title{Calculate by feature-pair mean from a CLR transformed approximation}
\usage{
getDeltaCLRMeans(x, y)
}
\arguments{
\item{x}{A vector of count data.}

\item{y}{A vector of count data.}
}
\description{
See rbeta.
}
