% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_params.R
\name{getDeltaCLRVars}
\alias{getDeltaCLRVars}
\title{Calculate by feature-pair variance from a CLR transformed approximation}
\usage{
getDeltaCLRVars(x, y)
}
\arguments{
\item{x}{A vector of count data.}

\item{y}{A vector of count data.}
}
\description{
See rbeta.
}
