% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_params.R
\name{getGenChiSqMean}
\alias{getGenChiSqMean}
\title{Calculate expected sum of squared differences from a CLR transformed approximation.}
\usage{
getGenChiSqMean(x, y)
}
\arguments{
\item{x}{A vector of count data.}

\item{y}{A vector of count data.}
}
\description{
See rbeta.
}
