% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta_params.R
\name{getGenChiSqVar}
\alias{getGenChiSqVar}
\title{Calculate variance of the expected sum of squared differences from a CLR transformed approximation.}
\usage{
getGenChiSqVar(x, y)
}
\arguments{
\item{x}{A vector of count data.}

\item{y}{A vector of count data.}
}
\description{
See rbeta.
}
