% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleuze.R
\name{getTableMeans}
\alias{getTableMeans}
\title{Estimate the approximated means from a CLR-transformed count sample.}
\usage{
getTableMeans(count_table, CLR_transformed = TRUE, cols_as_features = FALSE)
}
\arguments{
\item{count_table}{A table of count data with rows as features and columns as samples.}

\item{CLR_transformed}{A boolean, whether to return centered log-ratio (CLR) transformed values.}

\item{cols_as_features}{A boolean. Toggles whether rows or columns are samples}
}
\description{
See rbeta.
}
