% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perturb_by_CLR}
\alias{perturb_by_CLR}
\title{Perturb a CLR-transformed vector}
\usage{
perturb_by_CLR(x, by, prop = 1)
}
\arguments{
\item{x}{A vector of CLR-transformed data.}

\item{by}{A numeric. How much aitchison distance to perturb the vector x by.}

\item{prop}{A proportion (0-1). What proportion of the features should be the perturbation be divided over.}
}
\value{
A perturbed vector
}
\description{
Randomly perturb a CLR-transformed vector by a given aitchison distance.
}
