% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perturb_by_count}
\alias{perturb_by_count}
\title{Perturb a count data vector}
\usage{
perturb_by_count(x, by, prop = 1)
}
\arguments{
\item{x}{A vector of count data.}

\item{by}{A numeric. How much aitchison distance to perturb the vector x by.}

\item{prop}{A proportion (0-1). What proportion of the features should be the perturbation be divided over.}
}
\value{
A perturbed vector
}
\description{
Randomly perturb a count data vector by a given aitchison distance.
}
