% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_dist.R
\name{pw_dist}
\alias{pw_dist}
\title{Compute pairwise distances while ignoring shared zeroes/lowest values.}
\usage{
pw_dist(x, upper = FALSE, diag = FALSE)
}
\arguments{
\item{x}{a numeric matrix}

\item{upper}{logical value indicating whether the upper triangle of the distance matrix should be printed by `print.dist`.}

\item{diag}{logical value indicating whether the diagonal of the distance matrix should be printed by `print.dist`.}
}
\description{
Works like `dist`.
}
