% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{sampleBetaBinom}
\alias{sampleBetaBinom}
\title{samples from the probability density function of a feature in count data.}
\usage{
sampleBetaBinom(samples, k, n)
}
\arguments{
\item{samples}{An integer. How many samples should be taken}

\item{k}{An integer. How many counts were observerd}

\item{n}{An integer. How many obserations were made}
}
\description{
See rbeta.
}
