% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{sampleEachBetaBinom}
\alias{sampleEachBetaBinom}
\title{samples from the probability density function of a feature in count data for each feature.}
\usage{
sampleEachBetaBinom(samples, count_sample)
}
\arguments{
\item{samples}{An integer. How many samples should be taken}

\item{count_sample}{A vector of count data.}
}
\description{
samples from the probability density function of a feature in count data for each feature.
}
