% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{sampleLongitudinal}
\alias{sampleLongitudinal}
\title{samples from the approximated probability density function of a longitudinal set of CLR-transformed count samples.}
\usage{
sampleLongitudinal(samples, count_table, cols_as_features = F, adjust = F)
}
\arguments{
\item{samples}{An integer. How many samples should be taken.}

\item{count_table}{A vector of count data.}

\item{cols_as_features}{A boolean. Toggles whether rows or columns are samples.}

\item{adjust}{A boolean. Whether to Apply shrinkage. Experimental.}
}
\value{
a matrix with colums as samples and rows as features.
}
\description{
samples from the approximated probability density function of a longitudinal set of CLR-transformed count samples.
}
