% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{sampleTableCLR}
\alias{sampleTableCLR}
\title{samples from the approximated probability density function of a CLR-transformed count sample.}
\usage{
sampleTableCLR(samples, count_table)
}
\arguments{
\item{samples}{An integer. How many samples should be taken}

\item{count_table}{A vector of count data.}
}
\description{
samples from the approximated probability density function of a CLR-transformed count sample.
}
