% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleuze.R
\name{sCLR}
\alias{sCLR}
\title{Give shrunk means from a CLR-transformed count sample.}
\usage{
sCLR(count_table, cols_as_features = F)
}
\arguments{
\item{count_table}{A table of count data with rows as features and columns as samples.}

\item{cols_as_features}{A boolean. Toggles whether rows or columns are samples.}
}
\description{
See rbeta.
}
