% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleuze.R
\name{volatility}
\alias{volatility}
\title{Estimate volatility from count data in Aitchision distance.}
\usage{
volatility(count_table, metadata, adjust = TRUE)
}
\arguments{
\item{count_table}{A table of count data with rows as features and columns as samples.}

\item{metadata}{A vector depicting which samples are from the same source or group.}

\item{adjust}{A boolean. Whether to Apply shrinkage. Experimental.}
}
\description{
Compute feature-wise variance over time for a group of measurements.
}
