% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{double_center}
\alias{double_center}
\title{Double center a table.}
\usage{
double_center(x, log.transform = FALSE)
}
\arguments{
\item{x}{a numeric table.}

\item{log.transform}{A boolean. Whether to return log-transformed values.}
}
\description{
To center: Divide each value by its corresponding row and column geometric mean.
}
