% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{sampleGeomMeam}
\alias{sampleGeomMeam}
\title{samples from the log() of the geometric mean.}
\usage{
sampleGeomMeam(samples, count_sample)
}
\arguments{
\item{samples}{An integer. How many samples should be taken}

\item{count_samples}{A vector of count data.}
}
\description{
samples from the log() of the geometric mean.
}
