% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{sampleGeomMeanApprox}
\alias{sampleGeomMeanApprox}
\title{samples from the approximated probability density function of the geometric mean of a count sample.}
\usage{
sampleGeomMeanApprox(samples, count_sample)
}
\arguments{
\item{samples}{An integer. How many samples should be taken}

\item{count_sample}{A vector of count data.}
}
\description{
samples from the approximated probability density function of the geometric mean of a count sample.
}
